Implemented a lot more of the level 1 BLAS functions.

+++ src/sasum.c
@@ -4,7 +4,7 @@ int sasum(size_t N, float* X, int INCX){
     int d = 0, ix = 0;
     float abs_value = 0.0;
     for(d = 0; d < N; d++){
-        abs_value += fabs(X[INCX]);
+        abs_value += fabsf(X[INCX]);
         ix += INCX;
     }
     return abs_value;
